#53 Refactor: Modify EasterEggUtils to display a dialog on non-rooted devices
This commit refactors the `EasterEggUtils` class.

Previously, `EasterEggUtils` was an object that tracked click counts to trigger an easter egg.

The new implementation of `EasterEggUtils` is a class that depends on `RootRepository`. It introduces a `showIfForcedNonRoot` method. This method checks if the device is rooted using `rootRepo.isRooted()`. If the device is not rooted, it displays an `AlertDialog` with the title "System UI Destroyed" and the message "App will exit in 3 seconds". After 3 seconds, the application closes.