#86 style: Ignore ScopedStorage for external storage permissions
This commit adds `tools:ignore="ScopedStorage"` to the `READ_EXTERNAL_STORAGE` and `WRITE_EXTERNAL_STORAGE` permissions in the `AndroidManifest.xml`.

This is done to acknowledge that the app is intentionally not using Scoped Storage for these permissions, likely due to requirements for broader file system access, potentially related to root functionality.

Signed-off-by: gustyxpower <gustiadityamuzaky08@gmail.com>