#104 Refactor: Optimize CPU realtime update frequency
This commit changes the update interval for `RealtimeCpuInfo`.

The `getCpuRealtime()` function in `SystemRepository.kt` now checks if the time since the `lastCpuRealtimeUpdate` is less than 2000 milliseconds (previously 5000 milliseconds) before returning the cached `cachedCpuRealtimeInfo`. This results in more frequent updates of the CPU's real-time information if requested within the new, shorter interval.

Signed-off-by: gustyxpower <gustiadityamuzaky08@gmail.com>