#138 feat: Add GpuControlCard to TuningScreen
This commit integrates the `GpuControlCard` into the `TuningScreen`.

The `GpuControlCard` is now displayed as the second item in the `TuningScreen`, after `CpuGovernorCard` and before `ThermalCard`. It receives the `TuningViewModel` and has the `blur` parameter set to `true`.

Padding for the `TuningScreen`'s content `Column` has also been adjusted, applying both horizontal and vertical padding of `16.dp` in addition to the Scaffold's `innerPadding`.

Signed-off-by: gustyxpower <gustiadityamuzaky08@gmail.com>