#146 refactor: Rename TerminalViewModel to MiscViewModel and create MiscScreen
This commit renames `TerminalViewModel.kt` to `MiscViewModel.kt` and removes the `exec` function from it.

Additionally, a new screen `MiscScreen.kt` has been created. This screen displays a placeholder message indicating that the feature is under development, along with a request for support. The text in this screen has a typing animation effect.

Signed-off-by: gustyxpower <gustiadityamuzaky08@gmail.com>